requirements.txt
fastapi
uvicorn
jinja2
python-multipart
line-bot-sdk
bs4
requests
psycopg2
datetime